ARES-2: Character controller
Added Player prefab along with Player and Controller components.
The Controller component is responsible for the movement of the
character and collision detection, which is accomplished via
a customisable number of Physics2D.OverlapBox calls, which simulates
continuous collision, instead of Rigidbody2D and Collider2Ds to allow
for greater customisability and flexibility of movement.